refactor(modal): components/(modal)로 되어있던 경로 components로 이동
