+++ components/Modal/index.tsx
@@ -2,11 +2,12 @@
 
 import { useAtomValue } from "jotai";
 import { modalContext } from "@/context/index";
+import { background } from "./style.css";
 
 const Modal = () => {
   const modal = useAtomValue(modalContext);
 
-  return <>{modal.component}</>;
+  return <div className={background}>{modal.component}</div>;
 };
 
 export default Modal;
